Use Long data type for user status counters

+++ src/main/java/com/detectlanguage/responses/StatusResponse.java
@@ -4,24 +4,24 @@ import java.util.Date;
 
 public class StatusResponse extends Response {
     private Date date;
-    private Integer requests;
-    private Integer bytes;
+    private Double requests;
+    private Double bytes;
     private String plan;
     private Date plan_expires;
-    private Integer daily_requests_limit;
-    private Integer daily_bytes_limit;
+    private Double daily_requests_limit;
+    private Double daily_bytes_limit;
     private String status;
 
     public Date getDate() {
         return date;
     }
 
-    public Integer getRequests() {
-        return requests;
+    public Long getRequests() {
+        return requests.longValue();
     }
 
-    public Integer getBytes() {
-        return bytes;
+    public Long getBytes() {
+        return bytes.longValue();
     }
 
     public String getPlan() {
@@ -32,12 +32,12 @@ public class StatusResponse extends Response {
         return plan_expires;
     }
 
-    public Integer getDailyRequestsLimit() {
-        return daily_requests_limit;
+    public Long getDailyRequestsLimit() {
+        return daily_requests_limit.longValue();
     }
 
-    public Integer getDailyBytesLimit() {
-        return daily_bytes_limit;
+    public Long getDailyBytesLimit() {
+        return daily_bytes_limit.longValue();
     }
 
     public String getStatus() {
